[Validation] Move EcalDigis sources in to plugins dir
