+++ Validation/EcalDigis/plugins/CollHandle.cpp
@@ -1,0 +1,1 @@
+#include "CollHandle.h"
