+++ Validation/EcalDigis/plugins/CollHandle.h
