+++ Validation/EcalDigis/plugins/EcalDigisValidation.h
@@ -37,7 +37,7 @@
 #include "DataFormats/EcalDigi/interface/EcalDigiCollections.h"
 #include "SimDataFormats/CrossingFrame/interface/CrossingFrame.h"
 #include "SimDataFormats/CrossingFrame/interface/MixCollection.h"
-#include <Validation/EcalDigis/interface/EcalDigisValidation.h>
+#include "EcalDigisValidation.h"
 #include "CalibCalorimetry/EcalTrivialCondModules/interface/EcalTrivialConditionRetriever.h"
 
 #include <iostream>
