+++ Validation/EcalDigis/plugins/EcalMixingModuleValidation.h
@@ -55,7 +55,7 @@
 #include "CondFormats/ESObjects/interface/ESPedestals.h"
 #include "CondFormats/DataRecord/interface/ESPedestalsRcd.h"
 #include "CondFormats/ESObjects/interface/ESIntercalibConstants.h"
-#include <Validation/EcalDigis/interface/EcalBarrelDigisValidation.h>
+#include "EcalBarrelDigisValidation.h"
 #include "CalibCalorimetry/EcalTrivialCondModules/interface/EcalTrivialConditionRetriever.h"
 #include "Geometry/Records/interface/CaloGeometryRecord.h"
