--- Validation/EcalDigis/src/CollHandle.cpp
@@ -1,1 +1,0 @@
-#include "Validation/EcalDigis/src/CollHandle.h"
